requirements.txt
#to install requirements run 
#pip install -r requirements.txt
#ensure that you're running virtual environment 

asgiref==3.9.1  
Django==5.2.6  
djangorestframework==3.16.1 
pip==25.2   
sqlparse==0.5.3
tzdata==2025.2